Fixed bugs in code and tests:
Changed claim() and release() in logged/lockedSharedObject into protected(from public) to match technote.
removed tests that had to do with manual claim/release.
added test for creation of Accessor and accessing them.
Added test for the passing of an accessor to another thread to test accessing it.
Changed some test scenarios.